+++ src/lib/lforms/lhc-form-utils.js
@@ -5,6 +5,8 @@
 import CommonUtils from "./lhc-common-utils.js";
 import {InternalUtil} from "./internal-utils.js";
 import * as htmlparser2 from "htmlparser2";
+import copy from "fast-copy";
+import parse from 'style-to-object';
 
 import itemControls from '../item-controls.json';
 
@@ -307,23 +309,25 @@ const FormUtils = {
     }
 
     if (fhirData) {
-      fhirVersion = this._requireValidFHIRVersion(fhirVersion, fhirData);
+      // make a copy of the fhirData so the original data are not modified.
+      let fhirDataCopy = copy(fhirData);
+      fhirVersion = this._requireValidFHIRVersion(fhirVersion, fhirDataCopy);
       var fhir = LForms.FHIR[fhirVersion];
-      switch (fhirData.resourceType) {
+      switch (fhirDataCopy.resourceType) {
         case "DiagnosticReport":
-          formData = fhir.DiagnosticReport.mergeDiagnosticReportToLForms(formData, fhirData);
+          formData = fhir.DiagnosticReport.mergeDiagnosticReportToLForms(formData, fhirDataCopy);
           formData.hasSavedData = true; // will be used to determine whether to update or save
           break;
         case "Bundle":
           // Bundle should contain DiagnosticReport
-          if (fhirData.type === "searchset" &&
-              fhirData.entry.find(ele => ele.resource.resourceType === "DiagnosticReport")) {
-            formData = fhir.DiagnosticReport.mergeDiagnosticReportToLForms(formData, fhirData);
+          if (fhirDataCopy.type === "searchset" &&
+            fhirDataCopy.entry.find(ele => ele.resource.resourceType === "DiagnosticReport")) {
+            formData = fhir.DiagnosticReport.mergeDiagnosticReportToLForms(formData, fhirDataCopy);
             formData.hasSavedData = true; // will be used to determine whether to update or save
           }
           break;
         case "QuestionnaireResponse":
-          formData = fhir.SDC.mergeQuestionnaireResponseToLForms(formData, fhirData);
+          formData = fhir.SDC.mergeQuestionnaireResponseToLForms(formData, fhirDataCopy);
           formData.hasSavedData = true; // will be used to determine whether to update or save
           break;
       }
@@ -1051,23 +1055,104 @@ const FormUtils = {
   },
 
 
+  /** Check if URLs in CSS have values that are not allowed by FHIR specs
+   * (Remote URLs are not allowed)
+   * @param {*} cssValue CSS value on the 'style' attribute of a DOM element
+   * @returns A array of URLs that are not allowed.
+   */
+  _hasForbiddenCssUrl: function(cssValue) {
+    // possible usage of url() in the cssValue
+    // Note: css property name is case-sensitive.
+    //
+    // //No url(\'https://example.com/images/myImg.jpg\');
+    // url("https://example.com/images/myImg.jpg");
+    // url(\"https://example.com/images/myImg.jpg\");
+    // url('https://example.com/images/myImg.jpg');
+    // url(https://example.com/images/myImg.jpg);
+    // url("data:image/jpg;base64,iRxVB0…");
+    // url(\"data:image/jpg;base64,iRxVB0…\");
+    // url('data:image/jpg;base64,iRxVB0…');
+    // url(data:image/jpg;base64,iRxVB0…);
+    // url(myImg.jpg);
+    // url(#IDofSVGpath);
+
+    // Any external URLs are not allowed, including http, https, file
+    // or other protocols.
+    // url("https://...");
+    // url(\"https://...");
+    // url('https://...');
+    // url(https://...);
+    // url("http://...");
+    // url(\"http://...");
+    // url('http://...');
+    // url(http://...);
+    // url("file://...");
+    // url(\"file://...");
+    // url('file://...');
+    // url(file://...);
+
+    // sample output:
+    //  css= "url(\"img_tree.gif\"), url('file://local.jpg'),url('paper.gif'),url(http://google.com), url(\"example_with_url_inside.gif\"),url('url(123.png)')"
+    //  [
+    //   "url(\"img_tree.gif\")",
+    //   "url('paper.gif')",
+    //   "url('file://local.jpg')",
+    //   "url(http://google.com)",
+    //   "url(\"example_with_url_inside.gif\")",
+    //   "url('url(123.png)"
+    //  ]
+
+    let forbiddenURLs = [];
+    const CSS_URL_REGEXP = /url\(\s*["']?(.*?)["']?\s*\)/g;
+    let matched = cssValue.match(CSS_URL_REGEXP)
+
+    const URL_PARAM_REGEXP = /^url\(\s*[\\"']?[A-Za-z0-9]*\:\/\//
+    if (matched) {
+      for(let i=0; i<matched.length; i++) {
+        let urlString = matched[i];
+        if (urlString.match(URL_PARAM_REGEXP)) {
+          forbiddenURLs.push(urlString)
+        }
+      }
+    }
+
+    return forbiddenURLs;
+  },
+
   /**
    * Check and return not allowed tags within the HTML version of the help text.
    * See https://build.fhir.org/ig/HL7/sdc/rendering.html and
    * https://hl7.org/fhir/R4/narrative.html for allowed subset of the HTML tags.
    * @param {*} htmlNarrative
-   * @return [{array}] an array of invalid tags and attibutes
+   * @return [{array}] an array of invalid tags and attributes
    */
   checkForInvalidHtmlTags: function(htmlNarrative) {
     let invalidTagsAttributes=[];
-    let notAllowedTags = ['html','head', 'body', 'ref', 'script', 'form', 'base', 'link', 'xlink', 'iframe', 'object'];
-    let deprecatedTags = ['acronym', 'applet', 'basefont', 'big', 'blink', 'center', 'dir', 'embed', 'font',
-        'frame', 'frameset', 'isindex', 'noframes', 'marquee', 'menu', 'plaintext', 's', 'strike', 'tt', 'u'];
-    const FORBID_TAGS = notAllowedTags.concat(deprecatedTags);
+    let forbiddenTags = ['html','head', 'body', 'ref', 'script', 'form', 'base', 'link', 'xlink', 'iframe', 'object'];
+    let deprecatedTags = ['applet', 'basefont', 'blink', 'center', 'dir', 'embed', 'font',
+        'frame', 'frameset', 'isindex', 'noframes', 'marquee', 'menu', 'plaintext', 's', 'strike', 'u'];
+    const FORBIDDEN_TAGS = forbiddenTags.concat(deprecatedTags);
     const ALLOWED_URI_REGEXP = /^(?:data:|#|\/)/i;
-    const FORBID_ATTR = ['style'];
+    const FORBIDDEN_ATTR = [];
+    // https://developer.mozilla.org/en-US/docs/Web/CSS/url_function
+    const CSS_PROPERTIES_WITH_URL = [
+      "background",
+      "background-image",
+      "border",
+      "border-image",
+      "border-image-source",
+      "content",
+      "cursor",
+      "filter",
+      "list-style",
+      "list-style-image",
+      "mask",
+      "mask-image",
+      "offset-path",
+      "clip-path"
+    ];
 
-    // Tags (not in the FORBID_TAGS list above) that could have a URL value.
+    // Tags (not in the FORBIDDEN_TAGS list above) that could have a URL value.
     // See https://stackoverflow.com/questions/2725156/complete-list-of-html-tag-attributes-which-have-a-url-value
     // TBD: A full url in 'cite' might not be invalid.
     const TAGS_WITH_URL = {
@@ -1094,10 +1179,11 @@ const FormUtils = {
       "source": "srcset"  //'srcset' has one URL when <source> is included in <picture> and <video> (where multiple <source> tags are used instead).
     }
 
+    let that = this;
     const parser = new htmlparser2.Parser({
       onopentag(name, attributes) {
         // check tags
-        FORBID_TAGS.forEach(tag => {
+        FORBIDDEN_TAGS.forEach(tag => {
           if (name.toLocaleLowerCase() === tag) {
             invalidTagsAttributes.push({"tag": tag});
           }
@@ -1107,18 +1193,18 @@ const FormUtils = {
           if (name.toLocaleLowerCase() === tag) {
             for (const [attr, value] of Object.entries(attributes)) {
               urlAttrs.forEach(urlAttr => {
-                if(attr.toLocaleLowerCase() === urlAttr && !value.match(ALLOWED_URI_REGEXP)) {
+                if(attr === urlAttr && !value.match(ALLOWED_URI_REGEXP)) {
                   invalidTagsAttributes.push({"tag": tag, "attribute": urlAttr});
                 }
               })
             }
           }
         };
-        // check attributes that chould have multiple URL values
+        // check attributes that could have multiple URL values
         for (const [tag, urlAttr] of Object.entries(TAGS_WITH_MULTIPLE_URLS_IN_ONE_ATTR)) {
           if (name.toLocaleLowerCase() === tag) {
             for (const [attr, value] of Object.entries(attributes)) {
-              if(attr.toLocaleLowerCase() === urlAttr) {
+              if(attr === urlAttr) {
                 let urlValues = value.split(",");
                 urlValues.forEach(urlValue => {
                   if (!urlValue.trim().match(ALLOWED_URI_REGEXP)) {
@@ -1129,17 +1215,35 @@ const FormUtils = {
             }
           }
         };
-        // check attributes
+        // check attributes (not FORBIDDEN_ATTR for now)
+        // for (const [attr, value] of Object.entries(attributes)) {
+        //   FORBIDDEN_ATTR.forEach(forbiddenAttr => {
+        //     if(attr === forbiddenAttr) {
+        //       invalidTagsAttributes.push({"tag": name.toLocaleLowerCase(), "attribute": forbiddenAttr });
+        //     }
+        //   });
+        // }
+
+        // check "style" attribute for URLs
         for (const [attr, value] of Object.entries(attributes)) {
-          FORBID_ATTR.forEach(forbidAttr => {
-            if(attr.toLocaleLowerCase() === forbidAttr) {
-              invalidTagsAttributes.push({"tag": name.toLocaleLowerCase(), "attribute": forbidAttr });
+          if(attr === "style") {
+            // parse the CSS string
+            let cssObj = parse(value);
+            for (const [cssProp, cssValue] of Object.entries(cssObj)) {
+              CSS_PROPERTIES_WITH_URL.forEach(styleProp => {
+                if (cssProp.toLocaleLowerCase() === styleProp) {
+                  let forbiddenURLs = that._hasForbiddenCssUrl(cssValue);
+                  forbiddenURLs.forEach(urlString => {
+                    invalidTagsAttributes.push({"tag": name.toLocaleLowerCase(),
+                      "attribute": "style", "cssPropertyValue": styleProp + " : " + urlString });
+
+                  })
+                }
+              })
             }
-          });
+          }
         }
-
       }
-
       // Do nothing on onclosetag(name, attributes) {}
     });
 
@@ -1147,6 +1251,17 @@ const FormUtils = {
     parser.end();
 
     return invalidTagsAttributes;
+  },
+
+
+  /**
+   * Escapes a string for safe use as an HTML attribute.
+   * @param val the string to be escaped
+   * @return the escaped version of val
+   */
+  escapeAttribute: function (val) {
+    return val.replace(/&/g, '&amp;').replace(/"/g, '&quot;').replace(/'/g,
+      '&#39;').replace(/</g, '&lt;').replace(/>/g, '&gt;');
   },
 
   /**