+++ CHANGELOG.md
@@ -2,9 +2,21 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
-## [36.6.2] 2024-11-18
+## [36.7.2] 2024-12-03
 ### Changed
 - Upgraded to nodejs version 22.
+
+## [36.7.1] 2024-12-2
+### Fixed
+- Updated checkForInvalidHtml to allow more tags/attributes.
+- Fixed a bug that help text and copyright text could be out of position
+  if they are too long.
+- Fixed a bug that the QuestionnaireResponse data could be modified 
+  when passed in to the mergeFHIRDataIntoLForms function.
+
+## [36.7.0] 2024-11-20
+### Added
+- Support for rendering-xhtml on answerOption.
 
 ## [36.6.1] 2024-10-24
 ### Fixed